Cleaned up unsplash component in Admin
no issue

- Fixed a bug where the unsplash button didn't hide if the integration
  was disabled.
- Cleaned up Unsplash in preparation for moving it to it's own package
  in future.
